selftests/bpf: Factor out load_path and defines from test_sock_addr
Factor out useful defines and load_path() from the test_sock_addr self
test to use them in sock_addr_kern.
